refactor: Move plotting recipes to package extensions (#16)
