+++ Project.toml
@@ -9,17 +9,26 @@ ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Optim = "429524aa-4258-5aef-a3af-852621145aeb"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+
+[weakdeps]
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+PythonPlot = "274fc56d-3b97-40fa-a1cd-1b4a50311bf9"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
+
+[extensions]
+OpenCALPHADPlotsExt = "Plots"
+OpenCALPHADPythonPlotExt = "PythonPlot"
+OpenCALPHADRecipesBaseExt = "RecipesBase"
 
 [compat]
 DifferentiationInterface = "0.7.14"
 ForwardDiff = "0.10"
 JSON = "1.4.0"
 Optim = "2.0.0"
-Plots = "1.41.4"
+Plots = "1.41"
+PythonPlot = "1"
 RecipesBase = "1"
 StaticArrays = "1"
 julia = "1.9"
